worker/requirements.txt
torch==2.5.1
torchvision==0.20.1
transformers==4.48.1
accelerate==1.3.0
safetensors==0.5.2
Pillow==11.1.0
boto3==1.35.92
rapidfuzz==3.11.0
